+++ registrators/icp_pointmatcher.h
@@ -44,7 +44,7 @@ class IcpUsingPointMatcher : public Interface<PointType> {
 
   void setInputSource(const PointCloudSourcePtr& cloud) override;
   void setInputTarget(const PointCloudTargetPtr& cloud) override;
-  bool align(const Eigen::Matrix4f& guess, Eigen::Matrix4f& result) override;
+  bool align(const Eigen::Matrix4d& guess, Eigen::Matrix4d& result) override;
 
  protected:
   void loadDefaultConfig();
